@@ -4,8 +4,8 @@
   "private": true,
   "dependencies": {
     "axios": "^1.7.2",
-    "react": "^16.5.2",
-    "react-dom": "^16.5.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-router-dom": "^6.25.1",
     "react-s-alert": "^1.4.1",
     "react-scripts": "^5.0.1"
