@@ -1,127 +1,109 @@
-import React, { Component } from 'react';
+import React, { useState, useEffect } from 'react';
 import './Login.css';
-import { GOOGLE_AUTH_URL, FACEBOOK_AUTH_URL, GITHUB_AUTH_URL, ACCESS_TOKEN } from '../../constants';
+import { GOOGLE_AUTH_URL, FACEBOOK_AUTH_URL, GITHUB_AUTH_URL, ACCESS_TOKEN, USER_EMAIL } from '../../constants';
 import { login } from '../../util/APIUtils';
-import { Link, Navigate, useNavigate } from 'react-router-dom'
+import { Link, useNavigate, useLocation } from 'react-router-dom';
 import fbLogo from '../../img/fb-logo.png';
 import googleLogo from '../../img/google-logo.png';
 import githubLogo from '../../img/github-logo.png';
 import Alert from 'react-s-alert';
 
-class Login extends Component {
-    componentDidMount() {
-        const state = this.props.location?.state;
+const Login = ({ authenticated, loadCurrentlyLoggedInUser }) => {
+    const navigate = useNavigate();
+    const location = useLocation();
+
+    useEffect(() => {
+        const state = location.state;
 
         if (state && state.error) {
             setTimeout(() => {
                 Alert.error(state.error, {
                     timeout: 5000
                 });
-                this.props.history.replace({
-                    pathname: this.props.location.pathname,
-                    state: {}
-                });
+                navigate(location.pathname, { replace: true, state: {} });
             }, 100);
         }
-    }
-    
-    render() {
-        if(this.props.authenticated) {
-            return <Navigate
-                to={{
-                pathname: "/",
-                state: { from: this.props.location }
-            }}/>;            
-        }
+    }, [location, navigate]);
 
-        return (
-            <div className="login-container">
-                <div className="login-content">
-                    <h1 className="login-title">Login to SpringSocial</h1>
-                    <SocialLogin />
-                    <div className="or-separator">
-                        <span className="or-text">OR</span>
-                    </div>
-                    <LoginForm loadCurrentlyLoggedInUser={this.props.loadCurrentlyLoggedInUser} {...this.props} />
-                    <span className="signup-link">New user? <Link to="/signup">Sign up!</Link></span>
+    if (authenticated) {
+        return <Navigate to={{ pathname: "/", state: { from: location } }} />;
+    }
+
+    return (
+        <div className="login-container">
+            <div className="login-content">
+                <h1 className="login-title">Login to SpringSocial</h1>
+                <SocialLogin />
+                <div className="or-separator">
+                    <span className="or-text">OR</span>
                 </div>
+                <LoginForm loadCurrentlyLoggedInUser={loadCurrentlyLoggedInUser} />
+                <span className="signup-link">New user? <Link to="/signup">Sign up!</Link></span>
             </div>
-        );
-    }
-}
+        </div>
+    );
+};
 
-class SocialLogin extends Component {
-    render() {
-        return (
-            <div className="social-login">
-                <a className="btn btn-block social-btn google" href={GOOGLE_AUTH_URL}>
-                    <img src={googleLogo} alt="Google" /> Log in with Google</a>
-                <a className="btn btn-block social-btn facebook" href={FACEBOOK_AUTH_URL}>
-                    <img src={fbLogo} alt="Facebook" /> Log in with Facebook</a>
-                <a className="btn btn-block social-btn github" href={GITHUB_AUTH_URL}>
-                    <img src={githubLogo} alt="Github" /> Log in with Github</a>
-            </div>
-        );
-    }
-}
+const SocialLogin = () => {
+    return (
+        <div className="social-login">
+            <a className="btn btn-block social-btn google" href={GOOGLE_AUTH_URL}>
+                <img src={googleLogo} alt="Google" /> Log in with Google
+            </a>
+            <a className="btn btn-block social-btn facebook" href={FACEBOOK_AUTH_URL}>
+                <img src={fbLogo} alt="Facebook" /> Log in with Facebook
+            </a>
+            <a className="btn btn-block social-btn github" href={GITHUB_AUTH_URL}>
+                <img src={githubLogo} alt="Github" /> Log in with Github
+            </a>
+        </div>
+    );
+};
 
+const LoginForm = ({ loadCurrentlyLoggedInUser }) => {
+    const [email, setEmail] = useState('');
+    const [password, setPassword] = useState('');
 
-class LoginForm extends Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            email: '',
-            password: ''
-        };
-        this.handleInputChange = this.handleInputChange.bind(this);
-        this.handleSubmit = this.handleSubmit.bind(this);
-    }
+    const handleInputChange = (event) => {
+        const { name, value } = event.target;
+        if (name === 'email') setEmail(value);
+        if (name === 'password') setPassword(value);
+    };
 
-    handleInputChange(event) {
-        const target = event.target;
-        const inputName = target.name;        
-        const inputValue = target.value;
+    const handleSubmit = (event) => {
+        event.preventDefault();
 
-        this.setState({
-            [inputName] : inputValue
-        });        
-    }
-
-    handleSubmit(event) {
-        event.preventDefault();   
-
-        const loginRequest = Object.assign({}, this.state);
+        const loginRequest = { email, password };
 
         login(loginRequest)
-        .then(response => {
-            localStorage.setItem(ACCESS_TOKEN, response.accessToken);
-            Alert.success("You're successfully logged in!");
-            this.props.loadCurrentlyLoggedInUser();
-            //window.location.reload();
-        }).catch(error => {
-            Alert.error((error && error.message) || 'Oops! Something went wrong. Please try again!');
-        });
-    }
-    
-    render() {
-        return (
-            <form onSubmit={this.handleSubmit}>
-                <div className="form-item">
-                    <input type="email" name="email" 
-                        className="form-control" placeholder="Email"
-                        value={this.state.email} onChange={this.handleInputChange} required/>
-                </div>
-                <div className="form-item">
-                    <input type="password" name="password" 
-                        className="form-control" placeholder="Password"
-                        value={this.state.password} onChange={this.handleInputChange} required/>
-                </div>
-                <div className="form-item">
-                    <button type="submit" className="btn btn-block btn-primary">Login</button>
-                </div>
-            </form>                    
-        );
-    }
-}
+            .then(response => {
+                localStorage.setItem(ACCESS_TOKEN, response.accessToken);
+                localStorage.setItem(USER_EMAIL, email);
+                Alert.success("You're successfully logged in!");
+                loadCurrentlyLoggedInUser();
+                //window.location.reload();
+            }).catch(error => {
+                Alert.error((error && error.message) || 'Oops! Something went wrong. Please try again!');
+            });
+    };
 
-export default Login
+    return (
+        <form onSubmit={handleSubmit}>
+            <div className="form-item">
+                <input type="email" name="email"
+                    className="form-control" placeholder="Email"
+                    value={email} onChange={handleInputChange} required />
+            </div>
+            <div className="form-item">
+                <input type="password" name="password"
+                    className="form-control" placeholder="Password"
+                    value={password} onChange={handleInputChange} required />
+            </div>
+            <div className="form-item">
+                <button type="submit" className="btn btn-block btn-primary">Login</button>
+            </div>
+        </form>
+    );
+};
+
+export default Login;
